Extract canPrompt helper to shared prompts module
